♻️ refactor: consolidate to workspace-only entry and relocate components
Deprecate separate gallery/videos pages, consolidate everything into workspace.
- Remove /gallery and /videos pages, redirect to workspace
- Relocate ImageGallery, ImageGallerySkeleton to workspace/
- Move DuckDecodeDialog to selection/ for shared use
- Move MediaSortControls to workspace/
- Remove unused BotTab (merged into BotBuilderTab)
- Remove unused YoutubeDownloader component
- Remove unused process-store
- Update imports and exports across affected files
