+++ runninghub-nextjs/docs/images-components-relocation-todos.md
@@ -1,0 +1,6 @@
+# Images Components Relocation TODOs
+
+- [x] Move `MediaSortControls` to `src/components/workspace/MediaSortControls.tsx` and update imports.
+- [x] Move `DuckDecodeDialog` to `src/components/selection/DuckDecodeDialog.tsx` and update imports.
+- [x] Remove `src/components/images/index.ts` and delete `src/components/images/` folder.
+- [ ] Run `npm run build` in `runninghub-nextjs/`.
