--- runninghub-nextjs/src/components/images/index.ts
@@ -1,3 +1,0 @@
-export { ImageGallery } from "./ImageGallery";
-export { ImageProcessConfig } from "./ImageProcessConfig";
-export { MediaSortControls } from "./MediaSortControls";
